Improvements to map offline area download

+++ www/app/map/OfflineMapOverviewCtrl.js
@@ -1,6 +1,6 @@
 ﻿angular
     .module('RegObs')
-    .controller('OfflineMapOverviewCtrl', function ($ionicPlatform, $ionicLoading, $state, OfflineMap, AppLogging, $scope, Utility) {
+    .controller('OfflineMapOverviewCtrl', function ($ionicPlatform, $ionicLoading, $state, OfflineMap, AppLogging, $scope, Utility, RegobsPopup) {
         var vm = this;
         vm.offlineAreas = [];
 
@@ -8,7 +8,7 @@
             $ionicLoading.show();
             OfflineMap.getOfflineAreas()
                .then(function (result) {
-                   AppLogging.log('Offline fetched' + JSON.stringify(result));
+                   AppLogging.log('Offline fetched ' + JSON.stringify(result));
                    vm.offlineAreas = result;
                }, function (error) {
                    AppLogging.log('Could not get offline area ' + JSON.stringify(error));
@@ -17,19 +17,30 @@
                });
         };
 
-        vm.downloadArea = function() {
-            $state.go('mapareadownload');
+        var showConfirm = function () {
+            return RegobsPopup.confirm('Slett offlinekart',
+                'Er du sikker på at du vil slette alle områder som er lastet ned offline?');
+        };
+
+        vm.offlineAreDetailsClick = function (area) {
+            $state.go('offlineareadetails', area);
         };
 
         vm.deleteMaps = function () {
-            $ionicLoading.show();
-            OfflineMap.deleteAllOfflineAreas()
-                .finally(load);
+            showConfirm()
+                .then(function (response) {
+                    if (response) {
+                        $ionicLoading.show();
+                        OfflineMap.deleteAllOfflineAreas()
+                            .finally(load);
+                    }
+                });
         };
 
-        vm.humanFileSize = function(bytes) {
+        vm.humanFileSize = function (bytes) {
             return Utility.humanFileSize(bytes, true);
         };
+
 
         $scope.$on('$ionicView.enter', load);
     });
